Fixed incorrect CALLBACK_USB_GetDescriptor() wIndex parameter type (thanks to matlo).

+++ Bootloaders/MassStorage/Descriptors.c
@@ -131,7 +131,7 @@ const USB_Descriptor_Configuration_t ConfigurationDescriptor =
  *  USB host.
  */
 uint16_t CALLBACK_USB_GetDescriptor(const uint16_t wValue,
-                                    const uint8_t wIndex,
+                                    const uint16_t wIndex,
                                     const void** const DescriptorAddress)
 {
 	const uint8_t DescriptorType = (wValue >> 8);
