+++ Projects/AVRISP-MKII/AVRISP-MKII.h
@@ -82,7 +82,7 @@
 		void EVENT_USB_Device_ConfigurationChanged(void);
 
 		uint16_t CALLBACK_USB_GetDescriptor(const uint16_t wValue,
-		                                    const uint8_t wIndex,
+		                                    const uint16_t wIndex,
 		                                    const void** const DescriptorAddress,
 		                                    uint8_t* const DescriptorMemorySpace)
 		                                    ATTR_WARN_UNUSED_RESULT ATTR_NON_NULL_PTR_ARG(3) ATTR_NON_NULL_PTR_ARG(4);
